enhancement: Introduced Singleton class for styling to avoid redundancy.
