+++ example/lib/main.dart
@@ -21,6 +21,11 @@ class DemoScreen extends StatelessWidget {
   DemoScreen({Key? key}) : super(key: key);
   final TextEditingController _controller = TextEditingController();
 
+  InputFormFieldConfig inputFormFieldConfig = InputFormFieldConfig(
+    setBottomMargin: 10,
+    setBorderType: BorderType.outlined,
+  );
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -33,7 +38,6 @@ class DemoScreen extends StatelessWidget {
               /// Basic
               InputFormField(
                 textEditingController: _controller,
-                bottomMargin: 10,
               ),
 
               /// Custom Top Label
@@ -43,10 +47,6 @@ class DemoScreen extends StatelessWidget {
                   padding: EdgeInsets.only(bottom: 5),
                   child: Text("Custom Label"),
                 ),
-                borderRadius: BorderRadius.zero,
-                fillColor: Colors.grey.shade300,
-                bottomMargin: 10,
-                borderType: BorderType.none,
               ),
 
               /// Standard - Outlined Border
@@ -54,7 +54,6 @@ class DemoScreen extends StatelessWidget {
                 textEditingController: _controller,
                 labelText: "Standard - Outlined Border",
                 hintText: "Hint Text",
-                bottomMargin: 10, // Optional
               ),
 
               /// No Border - Filled
@@ -62,9 +61,6 @@ class DemoScreen extends StatelessWidget {
                 textEditingController: _controller,
                 labelText: "No Border - Filled",
                 hintText: "Hint Text",
-                fillColor: Colors.black12,
-                borderType: BorderType.none,
-                bottomMargin: 10, // Optional
               ),
 
               /// Default Password Field
@@ -73,7 +69,6 @@ class DemoScreen extends StatelessWidget {
                 labelText: "Default Password",
                 hintText: "*****",
                 password: EnabledPassword(),
-                bottomMargin: 10, // Optional
               ),
 
               /// Customize Default Password Field
@@ -81,7 +76,6 @@ class DemoScreen extends StatelessWidget {
                 textEditingController: _controller,
                 labelText: "Custom Default Password",
                 hintText: "*****",
-                obscuringCharacter: "*",
                 password: EnabledPassword(
                   showPasswordIcon: const Icon(
                     Icons.add,
@@ -92,19 +86,13 @@ class DemoScreen extends StatelessWidget {
                     color: Colors.amber,
                   ),
                 ),
-                bottomMargin: 10, // Optional
               ),
 
               /// Custom
               InputFormField(
                 textEditingController: _controller,
                 labelText: "Custom",
-                labelTextStyle: const TextStyle(fontSize: 25),
                 hintText: "Hint Text",
-                hintTextStyle: const TextStyle(fontSize: 18),
-                borderColor: Colors.deepOrange,
-                borderRadius: BorderRadius.circular(10),
-                bottomMargin: 10,
                 // Optional
                 floatingLabelBehavior: FloatingLabelBehavior.always,
               ),
@@ -113,14 +101,7 @@ class DemoScreen extends StatelessWidget {
               InputFormField(
                 textEditingController: _controller,
                 labelText: "Custom 2",
-                labelTextStyle: const TextStyle(color: Colors.deepOrange),
                 hintText: "Hint Text",
-                hintTextStyle: const TextStyle(color: Colors.deepOrangeAccent),
-                borderType: BorderType.outlined,
-                borderColor: Colors.deepOrange,
-                bottomMargin: 10,
-                // Optional
-                borderRadius: BorderRadius.circular(10),
               ),
 
               /// Custom 3 (Suffix)
@@ -129,8 +110,6 @@ class DemoScreen extends StatelessWidget {
                 prefix: const Icon(Icons.mail),
                 labelText: "Email",
                 hintText: "abc@email.com",
-                borderType: BorderType.outlined,
-                bottomMargin: 10, // Optional
               ),
 
               /// Custom 4 (Suffix, Prefix)
@@ -140,8 +119,6 @@ class DemoScreen extends StatelessWidget {
                 labelText: "Email",
                 hintText: "abc@email.com",
                 suffix: const Icon(Icons.info),
-                borderType: BorderType.outlined,
-                bottomMargin: 10, // Optional
               ),
             ],
           ),
