+++ example/pubspec.yaml
@@ -30,6 +30,7 @@ dependencies:
   flutter:
     sdk: flutter
   input_form_field:
+    path: ../
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
